packages/tools/skill-0-youtube-fetcher/requirements.txt
google-api-python-client>=2.0.0
google-genai>=1.0.0
supabase>=2.0.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
